Allow typed composite constructors to accept scalar literals
Typed constructors like vec3d(), mat22d(), quatd(), and transformd() now
accept float and int literals without requiring explicit casts.  For
example, wp.vec3d(1.00000005, 2.0, 3.0) preserves double precision of
the literals, and wp.vec3h(1.0, 2.0, 3.0) implicitly converts to half.

Previously Var.__init__ canonicalized Python float literals to float32,
causing scalar_infer_type to infer float32 for the constructor dtype,
which then conflicted with the explicit dtype (e.g. float64 for vec3d).

The fix has two parts per constructor type:
- value_func: when dtype is explicitly provided and all scalar arguments
  are compile-time constants, accept a type mismatch instead of raising.
  Runtime variables with mismatched types still error at codegen time.
- dispatch_func: cast constant scalar Vars to the target dtype so the
  emitted C++ declarations use full precision.

+++ warp/tests/matrix/test_mat_lite.py
@@ -15,10 +15,10 @@ def test_matrix_constructor_value_func():
     b = wp.matrix(a, shape=(2, 2))
     c = mat32d()
     d = mat32d(c, shape=(3, 2))
-    e = mat32d(wp.float64(1.0), wp.float64(2.0), wp.float64(1.0), wp.float64(2.0), wp.float64(1.0), wp.float64(2.0))
+    e = mat32d(1.0, 2.0, 1.0, 2.0, 1.0, 2.0)
     f = wp.matrix_from_rows(
-        wp.vec3d(wp.float64(1.0), wp.float64(2.0), wp.float64(3.0)),
-        wp.vec3d(wp.float64(1.0), wp.float64(2.0), wp.float64(3.0)),
+        wp.vec3d(1.0, 2.0, 3.0),
+        wp.vec3d(1.0, 2.0, 3.0),
     )
     g = wp.types.matrix(1.0, shape=(3, 2))
 
@@ -33,7 +33,7 @@ def test_constructors_explicit_precision():
     # construction for custom matrix types
     eye = wp.identity(dtype=wp.float16, n=2)
     zeros = wp.types.matrix(shape=(2, 2), dtype=wp.float16)
-    custom = wp.types.matrix(wp.float16(0.0), wp.float16(1.0), wp.float16(2.0), wp.float16(3.0), shape=(2, 2))
+    custom = wp.types.matrix(0.0, 1.0, 2.0, 3.0, shape=(2, 2), dtype=wp.float16)
 
     for i in range(2):
         for j in range(2):
